Export EmphasizeElementProps from common, deprecate from frontend.
