+++ common/api/summary/imodeljs-common.exports.csv
@@ -9,6 +9,7 @@ public;AmbientLightProps
 public;AmbientOcclusion
 alpha;AnalysisStyle
 alpha;AnalysisStyleProps
+public;AppearanceOverrideProps
 public;AreaFillProps
 public;AreaPattern
 public;AuxCoordSystem2dProps 
@@ -155,6 +156,7 @@ public;ElementIdsAndRangesProps
 public;ElementLoadOptions
 public;ElementLoadProps 
 public;ElementProps 
+public;EmphasizeElementsProps
 beta;EntityMetaData 
 beta;EntityMetaDataProps
 public;EntityProps
@@ -175,6 +177,7 @@ internal;FeatureGates
 internal;FeatureIndex
 internal;FeatureIndexType
 public;FeatureOverrides 
+public;FeatureOverrideType
 public;FeatureTable 
 internal;FeatureTableHeader
 public;FilePropertyProps
