+++ common/api/summary/imodeljs-frontend.exports.csv
@@ -32,7 +32,8 @@ internal;AnimatedTreeReference
 internal;AnimationBranchState
 internal;AnimationBranchStates = Map
 public;Animator
-public;AppearanceOverrideProps
+public;AppearanceOverrideProps = AppearanceOverrideProps_2
+deprecated;AppearanceOverrideProps = AppearanceOverrideProps_2
 internal;ArcGisErrorCode
 internal;ArcGisGenerateTokenOptions
 internal;ArcGISMapLayerImageryProvider 
@@ -147,7 +148,8 @@ alpha;class ElementSetTool
 public;ElementState 
 internal;EllipsoidTerrainProvider 
 public;EmphasizeElements 
-public;EmphasizeElementsProps
+public;EmphasizeElementsProps = EmphasizeElementsProps_2
+deprecated;EmphasizeElementsProps = EmphasizeElementsProps_2
 beta;EngineeringLengthDescription 
 public;EntityState 
 public;Environment
@@ -167,7 +169,6 @@ internal;eyeToCartographicOnGlobeFromGcs(viewport: ScreenViewport, preserveHeigh
 alpha;FeatureLogBatchClient
 alpha;FeatureLogBatchOptions
 public;FeatureOverrideProvider
-public;FeatureOverrideType
 public;FeatureSymbology
 alpha;FeatureTrackingManager
 alpha;FeatureTrackingProps
