+++ common/changes/@bentley/imodeljs-frontend/emphasize-element-props_2021-06-03-14-31.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/imodeljs-frontend",
+      "comment": "Export EmphasizeElementProps from common, deprecate from frontend.",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/imodeljs-frontend",
+  "email": "65233531+bbastings@users.noreply.github.com"
+}
